comply with Google Analytics 4 (GA4)
In 2023-07-01, Universal Analytics (UA), Google Analytics 3
and below stopped working in favour of GA4. Unfortunately,
GA4 is completely incompatible with them, offering new
endpoints, methods, dimensions and metrics to query.

This means that the code that interfaces with that API also
needs to change in backwards incompatible ways.

This is the basic implementation that adapts core API
functionality.